ABC Notation is working with a single test file grabbed from internet.  More extensive tests pending.

+++ EightBitTunes/TuneManager.h
@@ -22,7 +22,7 @@ class TuneManager {
     int tuneFreq[MAX_NOTE_BUFFER];
     int tuneDur[MAX_NOTE_BUFFER];
     char** sounds;
-    void addNotesToTune(byte numOfNotesToAdd);
+    void addNotesToTune(int numOfNotesToAdd);
 };
 
 #endif
